Add git_info.py and read_update_rate.py Update dsr.srdf: rename all group/link names to match model name
