--- dsr_controller2/config/dsr_update_rate.yaml
@@ -1,3 +1,0 @@
-controller_manager:
-  ros__parameters:
-    update_rate: 100  # Hz
